feat: repository added

+++ src/Lab2/PersonalComputerConfigurator/Models/RamCharacterisics/FrequencyVoltagePair.cs
@@ -1,3 +1,3 @@
 namespace Itmo.ObjectOrientedProgramming.Lab2.PersonalComputerConfigurator.Models.RamCharacterisics;
 
-public record FrequencyVoltagePair(int Mhz, int V);
+public record FrequencyVoltagePair(int Mhz, double V);
