+++ src/Lab2/PersonalComputerConfigurator/Models/XmpProfileCharacteristics/Voltage.cs
@@ -1,3 +1,3 @@
 namespace Itmo.ObjectOrientedProgramming.Lab2.PersonalComputerConfigurator.Models.XmpProfileCharacteristics;
 
-public record Voltage(int V);
+public record Voltage(double V);
